+++ src/tt.h
@@ -86,7 +86,7 @@ public:
   void new_search() { generation8 += 4; } // Lower 2 bits are used by Bound
   uint8_t generation() const { return generation8; }
   TTEntry* probe(const Key key, bool& found) const;
-  void resize(uint64_t mbSize);
+  void resize(size_t mbSize);
   void clear();
 
   // The lowest order bits of the key are used to get the index of the cluster
